pyproject.toml
[tool.poetry]
name = "eleina-agent"
version = "0.1.0"
description = "Anime character intelligent agent"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.1.0"
requests = "^2.32.0"
numpy = "^1.26.4"
scikit-learn = "^1.4.0"
PyPDF2 = "^3.0.1"
pdfminer-six = "^20231228"
python-docx = "^1.0.0"
httpx = "^0.28.0"
redis = "^7.4.0"
sqlalchemy = "^2.0.49"
pymysql = "^1.1.0"
pymilvus = "^2.4.0"
rank-bm25 = "^0.2.2"

# ✅✅✅ 官方真实存在 + 100% 兼容 ✅✅✅
langchain = "^0.2.10"
langchain-core = "^0.2.36"
langchain-community = "^0.2.10"
langgraph = "^0.2.50"
langchain-milvus = "^0.1.3"
dashscope = "^1.25.16"
pillow = "^12.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.4.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"